requirements.txt
pandas==1.5.3
git+https://github.com/turkish-gekko/service-binance-rest@main#egg=turkish_gekko_binance_service
ta==0.9.0
# SYSTEM_VERSION_COMPAT=0 pip install tensorflow-macos tensorflow-metal
python-dotenv
pandas_ta